refactor: 구현 패키지와 인터페이스 패키지 분리
