+++ src/main/java/me/namuhuchutong/ladder/domain/implement/Results.java
@@ -1,6 +1,6 @@
-package me.namuhuchutong.ladder.domain;
+package me.namuhuchutong.ladder.domain.implement;
 
-import me.namuhuchutong.ladder.domain.wrapper.Result;
+import me.namuhuchutong.ladder.domain.implement.wrapper.Result;
 
 import java.util.Arrays;
 import java.util.List;
