+++ src/test/java/me/namuhuchutong/ladder/NameAndResultTest.java
@@ -1,8 +1,8 @@
 package me.namuhuchutong.ladder;
 
 import me.namuhuchutong.ladder.dto.NameAndResult;
-import me.namuhuchutong.ladder.domain.wrapper.Name;
-import me.namuhuchutong.ladder.domain.wrapper.Result;
+import me.namuhuchutong.ladder.domain.implement.wrapper.Name;
+import me.namuhuchutong.ladder.domain.implement.wrapper.Result;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
