+++ src/test/java/me/namuhuchutong/ladder/RowsTest.java
@@ -1,12 +1,12 @@
 package me.namuhuchutong.ladder;
 
-import me.namuhuchutong.ladder.domain.Rows;
+import me.namuhuchutong.ladder.domain.implement.Rows;
 import me.namuhuchutong.ladder.domain.engine.ScaffoldFactory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import static me.namuhuchutong.ladder.domain.wrapper.LadderExpression.*;
+import static me.namuhuchutong.ladder.domain.implement.wrapper.LadderExpression.*;
 import static org.assertj.core.api.Assertions.*;
 
 class RowsTest {
